+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/container.py
@@ -462,7 +462,7 @@ class ContainerProxy(object):
         """
         request_options = _build_options(kwargs)
         response_hook = kwargs.pop('response_hook', None)
-        request_options["disableIdGeneration"] = True
+        request_options["disableAutomaticIdGeneration"] = True
         pre_trigger_include = kwargs.pop('pre_trigger_include', None)
         if pre_trigger_include is not None:
             request_options["preTriggerInclude"] = pre_trigger_include
@@ -512,7 +512,7 @@ class ContainerProxy(object):
         item_link = self._get_document_link(item)
         request_options = _build_options(kwargs)
         response_hook = kwargs.pop('response_hook', None)
-        request_options["disableIdGeneration"] = True
+        request_options["disableAutomaticIdGeneration"] = True
         pre_trigger_include = kwargs.pop('pre_trigger_include', None)
         if pre_trigger_include is not None:
             request_options["preTriggerInclude"] = pre_trigger_include