+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_base_client.py
@@ -20,7 +20,7 @@ class TextAnalyticsApiVersion(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Text Analytics API versions supported by this package"""
 
     #: this is the default version
-    V3_2_PREVIEW = "v3.2-preview.2"
+    V2022_03_01_PREVIEW = "2022-03-01-preview"
     V3_1 = "v3.1"
     V3_0 = "v3.0"