+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_extract_key_phrases.py
@@ -6,7 +6,7 @@
 import pytest
 import platform
 import functools
-
+import json
 from azure.core.exceptions import HttpResponseError, ClientAuthenticationError
 from azure.core.credentials import AzureKeyCredential
 from testcase import TextAnalyticsTest, TextAnalyticsPreparer
@@ -549,19 +549,23 @@ class TestExtractKeyPhrases(TextAnalyticsTest):
         assert res == "cls result"
 
     @TextAnalyticsPreparer()
-    @TextAnalyticsClientPreparer(client_kwargs={"api_version": TextAnalyticsApiVersion.V3_0})
-    @recorded_by_proxy
-    def test_string_index_type_not_fail_v3(self, client):
-        # make sure that the addition of the string_index_type kwarg for v3.1-preview.1 doesn't
-        # cause v3.0 calls to fail
-        client.extract_key_phrases(["please don't fail"])
-
-    @TextAnalyticsPreparer()
-    @TextAnalyticsClientPreparer()
+    @TextAnalyticsClientPreparer(client_kwargs={"api_version": TextAnalyticsApiVersion.V3_1})
     @recorded_by_proxy
     def test_disable_service_logs(self, client):
         def callback(resp):
             assert resp.http_request.query['loggingOptOut']
+        client.extract_key_phrases(
+            documents=["Test for logging disable"],
+            disable_service_logs=True,
+            raw_response_hook=callback,
+        )
+
+    @TextAnalyticsPreparer()
+    @TextAnalyticsClientPreparer(client_kwargs={"api_version": TextAnalyticsApiVersion.V2022_03_01_PREVIEW})
+    @recorded_by_proxy
+    def test_disable_service_logs_body_param(self, client):
+        def callback(resp):
+            assert json.loads(resp.http_request.body)['parameters']['loggingOptOut']
         client.extract_key_phrases(
             documents=["Test for logging disable"],
             disable_service_logs=True,